Revert "adds validations for get by ids endpoints"
This reverts commit cb9ea5326b46ea10843a94264e11525ebdfb49a9.

+++ src/main/java/com/featuredoc/models/UserRoleId.java
@@ -12,14 +12,14 @@ import lombok.*;
 @AllArgsConstructor
 public class UserRoleId implements Serializable {
 
-    private Long roleID;
-    private Long userID;
+    private Integer roleID;
+    private Integer userID;
 
-    public Long getUserID() {
+    public Integer getUserID() {
         return userID;
     }
 
-    public Long getRoleID() {
+    public Integer getRoleID() {
         return roleID;
     }
